requirements.txt
ai2thor
opencv-python
numpy
tqdm
transformers
torch
pyyaml
prior
